[ modify ] 1.新增example.env 2.content 去除空格trim() 3.errorHandle 將 error 加入，取得 errorMessages 4.POST 或 PATCH : content內容不能為空， 5.Delete 或 PATCH是否有此id貼文 與 是否是有效的 ObjectId 6.PATCH 更新貼文成功時，在第三個變數帶入 { new: true } 取得最新的更新後的資訊。

+++ server.js
@@ -33,18 +33,23 @@ const requestListener = async(req, res) => {
     req.on('end', async () => {
       try {
         const data = JSON.parse(body)
+        const { name, tags, type, image, content } = data
+        
+        if(!content?.trim()) throw new Error('貼文內容不能空白ㄛ~');
+        
         const newPosts = await Posts.create(
           {
-            name: data.name,
-            tags: data.tags,
-            type: data.type,
-            image: data.image,
-            content: data.content
+            name,
+            tags,
+            type,
+            image,
+            content: content.trim()
           }
         )
         successHandle(res, newPosts)
+        
       } catch (error) {
-        errorHandle(res)
+        errorHandle({ res, error })
       } 
     })
   } else if ( req.url === '/posts' && req.method === 'DELETE') {
@@ -55,22 +60,51 @@ const requestListener = async(req, res) => {
     // 刪除一則貼文
     try {
       const id = req.url.split('/').pop()
+
+      // 檢查 id 是否是有效的 ObjectId
+      if (!mongoose.Types.ObjectId.isValid(id))  throw new Error(`無效的貼文 ID : ${id}`);
+
+      // 貼文 是否存在
+      const post = await Posts.findById(id);
+      if (!post) throw new Error(`此貼文不存在：${id}`);
+
       await Posts.findByIdAndDelete(id) // 刪除
       const postData = await Posts.find()
       successHandle(res, postData)
+      
     } catch (error) {
-      errorHandle(res)
+      errorHandle({ res, error })
     }
   } else if ( req.url.startsWith('/posts/') && req.method === 'PATCH') {
     req.on('end' , async () => {
       try {
         const id = req.url.split('/').pop()
-        const updateData = JSON.parse(body)
-        await Posts.findByIdAndUpdate(id, updateData)
-        const postData = await Posts.find()
-        successHandle(res, postData)
+
+        // 檢查 id 是否是有效的 ObjectId
+        if (!mongoose.Types.ObjectId.isValid(id))  throw new Error(`無效的貼文 ID : ${id}`);
+
+        // 貼文 是否存在
+        const post = await Posts.findById(id);
+        if (!post) throw new Error(`此貼文不存在：${id}`);
+        
+        const data = JSON.parse(body)
+        const { name, tags, type, image, content } = data
+        if( !content?.trim()) throw new Error('貼文內容不能空白ㄛ~');
+        
+        const updateData = {
+          name,
+          tags,
+          type,
+          image,
+          content: content.trim()
+        }
+
+        // Patch 更新貼文成功時，可以在第三個變數帶入 { new: true } 就能取得最新的更新後的資訊。
+        const updatedPost = await Posts.findByIdAndUpdate(id, updateData, { new: true });
+        successHandle(res, updatedPost)      
+        
       } catch (error) {
-        errorHandle(res)
+        errorHandle({ res, error })
       }
     })
   } else if ( req.method === 'OPTIONS') {
@@ -84,7 +118,6 @@ const requestListener = async(req, res) => {
     }))
     res.end()
   }
-
 }
 
 const server = http.createServer(requestListener)
